Refactor r2d2_utils
