+++ src/swell/tasks/get_background.py
@@ -9,7 +9,7 @@
 
 
 from swell.tasks.base.task_base import taskBase
-from swell.utilities.r2d2 import load_r2d2_credentials, get_r2d2_model_name
+from swell.utilities.r2d2_utils import load_r2d2_credentials, get_r2d2_model_name
 
 import isodate
 import os
