+++ src/swell/tasks/get_geovals.py
@@ -11,7 +11,7 @@
 import os
 
 from swell.tasks.base.task_base import taskBase
-from swell.utilities.r2d2 import load_r2d2_credentials
+from swell.utilities.r2d2_utils import load_r2d2_credentials
 from r2d2 import fetch
