+++ src/swell/tasks/get_ncdiags.py
@@ -10,7 +10,7 @@
 import os
 from swell.tasks.base.task_base import taskBase
 from r2d2 import fetch
-from swell.utilities.r2d2 import load_r2d2_credentials
+from swell.utilities.r2d2_utils import load_r2d2_credentials
 
 # --------------------------------------------------------------------------------------------------
