+++ src/swell/tasks/save_forecast.py
@@ -16,7 +16,7 @@ from r2d2 import store
 
 from swell.tasks.base.task_base import taskBase
 from swell.utilities.datetime_util import datetime_formats
-from swell.utilities.r2d2 import load_r2d2_credentials
+from swell.utilities.r2d2_utils import load_r2d2_credentials
 
 
 # --------------------------------------------------------------------------------------------------
